requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==6.0.1
python-dotenv==1.1.1
SQLAlchemy-serializer==1.4.22

